utils/session.py
from utils.constants import upwork_login_url, cloudfare_challenge_div_id, upwork_url, home_url

from nyx.page import NyxPage

from httpx import AsyncClient
from pydantic import BaseModel
import asyncio

class Session:
    def __init__(self, task_id:int, page:NyxPage, username: str, password: str, security_answer: str = None, status_endpoint:str = None, payload_endpoint:str = None, payload:BaseModel = None):
        self.task_id = task_id
        self.username = username
        self.password = password
        self.security_answer = security_answer
        self.client = None
        self.page = page
        self.payload:BaseModel = payload
        self.status_endpoint = status_endpoint
        self.payload_endpoint = payload_endpoint
        self.status = {}
    
    async def setup_client(self):
        try:
            self.client = AsyncClient()
            return True 
        except Exception as e:
            self.update_status("Failed", f"Error setting up HTTP client: {e}")
            # self.send_status()
            self.print_status()
            return False
        
    async def close_client(self):
        if self.client:
            await self.client.aclose()
            
    def update_status(self, status:str, message:str):
        self.status["status"] = status
        self.status["message"] = message
    async def logout(self):
        try:
            await self.page.click('button[aria-describedby="options-theme-popover"]')
            await asyncio.sleep(0.5)
            await self.page.click('button[data-cy="logout-trigger"]')
            await asyncio.sleep(2)
        except Exception as e:
            raise e
                    
    async def login(self, remember_me:bool = True, to_scrape:bool = False):
        try:
            await self.page.goto(upwork_login_url,captcha_selector=cloudfare_challenge_div_id,wait_until= "domcontentloaded",referer=upwork_url) 
            logged_in = await self.page.check_for_element('section[data-test="freelancer-sidebar-profile"]')
            if logged_in:
                print("Already logged in")
                logged_in_user = await self.page.get_text_content('a.profile-title')
                print(f"Logged in user: {logged_in_user}")
                if "Maharuf" in logged_in_user:
                    self.update_status("Success", "Changing accounts - Logging out first")
                    self.print_status()
                    await self.logout()
            login_page = await self.page.check_for_element("#login_username")
            await asyncio.sleep(2)
            if login_page:
                await self.page.fill_field_and_enter('#login_username', self.username)
                await asyncio.sleep(3)
                await self.page.wait_for_element('#login_password')
                if remember_me:
                    await self.page.click('#login_rememberme')
                await self.page.fill_field_and_enter('#login_password', self.password)
                await asyncio.sleep(3)
                if self.security_answer and await self.page.check_for_element('#login_answer'):
                    await self.page.fill_field_and_enter('#login_answer', self.security_answer)
            elif await self.page.check_for_element('section[data-test="freelancer-sidebar-profile"]'):
                self.update_status("Success", "Already logged in")
                return True
            else:
                # await self.send_status("Failed", "Login page not found")
                self.print_status()
                return False
        except Exception as e:
            # await self.send_status("Failed", f"Error during login: {e}")
            self.print_status()
            await self.page.goto(home_url)
            return False
        finally:
            if not to_scrape:
                await self.page.goto(home_url)
                
    async def send_status(self, status: str = None, message: str = None):
        if not self.status_endpoint:
            self.status["status"] = "Failed"
            self.status["message"] = "Set the status_endpoint parameter in Session initialisation."
            return False
        if not self.client:
            await self.setup_client()
        try:
            if status and message:
                self.update_status(status, message)
            await self.client.post(self.status_endpoint, json=self.status)
            return True
        except Exception as e:
            self.status["status"] = "Failed"
            self.status["message"] = f"Error sending status: {e}"
            return False
        
    def print_status(self):
        print(f"{self.status["status"]} -- {self.status["message"]}")
        
    async def send_payload(self):
        if not self.payload_endpoint:
            self.status["status"] = "Failed"
            self.status["message"] = "Set the payload_endpoint parameter in Session initialisation."
            return False
        if not self.client:
            await self.setup_client()
        try:
            print(self.payload.model_dump_json())
            await self.client.post(self.payload_endpoint, json=self.payload.model_dump())
            return True
        except Exception as e:
            self.status["status"] = "Failed"
            self.status["message"] = f"Error sending payload: {e}"
            print(self.status)
            return False

        