requirements.txt
asyncpg==0.30.0
dotenv==0.9.9
fastapi==0.116.1
httpx==0.28.1
joblib
langchain==0.3.27
langchain-community==0.3.29
langchain-core==0.3.76
langchain-openai==0.3.33
langchain-postgres==0.0.15
langchain-text-splitters==0.3.11
langgraph==0.6.8
langgraph-checkpoint==2.1.1
langgraph-checkpoint-postgres==2.0.23
matplotlib==3.10.3
numpy==2.3.1
openai==1.107.1
openpyxl==3.1.5
pgvector==0.3.6
pandas==2.3.1
playwright==1.53.0
playwright-stealth==2.0.0
psycopg2-binary==2.9.9
psycopg-pool==3.2.6
python-dotenv==1.1.1
python-ghost-cursor==0.1.1
uvicorn==0.35.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-jose==3.5.0
alembic==1.18.4
